packages/data_prep_cataloguer/requirements.txt
google-cloud-storage==1.17
PyYAML==5.3.1
requests==2.23.0
tqdm==4.46.0
typing-extensions==3.7.4.2
urllib3==1.25.9
coverage
flake8
pg8000==1.13.2
SQLAlchemy==1.3.17
